Switch from NPM to Yarn
Lockfiles are great. We should use them whenever possible.
Unfortunately, due to an unresolved issue that is apparently difficult
to reproduce, the presence of package-lock.json occasionally causes npm
not to install some dependencies.

This hasn't been an issue on this project yet, but it has caused several
Heroku builds to fail on other projects.

More information:
https://github.com/npm/npm/issues/16839